refactor BufferedIndexReplays into separate module, to avoid circular dependency between IndexStorageInfo and UnboundIndex for serialization code
